add linkto componennt and update card and gallery to use it
