--- components/card.module.css
@@ -1,4 +1,0 @@
-.aLike {
-    position: relative;
-    overflow: hidden;
-}
